requirements.txt
line-bot-sdk
Flask
openai
python-dotenv